system/pyproject.toml
[tool.poetry]
name = "system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
python-socketio = {extras = ["asyncio_client"], version = "^5.0.4"}
fastapi-socketio = "^0.0.6"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
fastapi-utils = "^0.2.1"
simplejson = "^3.17.2"
pyserial = "^3.5"
pytest-asyncio = "^0.14.0"
websocket_client = "^0.57.0"
vectormath = "^0.2.2"
numpy = "^1.20.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-watch = "^4.2.0"
pytest-flakefinder = "^1.0.0"
pytest-profiling = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
